+++ package.json
@@ -48,7 +48,6 @@
     "addons-scanner-utils": "6.1.0",
     "ajv": "6.12.6",
     "ajv-merge-patch": "4.1.0",
-    "async": "3.2.2",
     "chalk": "4.1.2",
     "cheerio": "1.0.0-rc.10",
     "columnify": "1.5.4",
@@ -64,12 +63,10 @@
     "image-size": "1.0.0",
     "is-mergeable-object": "1.1.1",
     "jed": "1.1.1",
-    "natural-compare-lite": "1.4.0",
     "os-locale": "5.0.0",
     "pino": "7.1.0",
     "postcss": "8.3.11",
     "relaxed-json": "1.0.3",
-    "request": "2.88.2",
     "semver": "7.3.5",
     "sha.js": "2.4.11",
     "source-map-support": "0.5.20",
@@ -79,6 +76,7 @@
     "yauzl": "2.10.0"
   },
   "devDependencies": {
+    "async": "3.2.2",
     "@babel/cli": "7.16.0",
     "@babel/core": "7.16.0",
     "@babel/eslint-parser": "7.16.0",
@@ -107,10 +105,12 @@
     "markdown-it": "12.2.0",
     "markdown-it-anchor": "8.4.1",
     "markdown-it-emoji": "2.0.0",
+    "natural-compare-lite": "1.4.0",
     "po2json": "1.0.0-beta-3",
     "prettier": "2.4.1",
     "pretty-quick": "3.1.1",
     "raw-loader": "4.0.2",
+    "request": "2.88.2",
     "shelljs": "0.8.4",
     "sinon": "12.0.1",
     "tar": "6.1.11",
