+++ src/main.js
@@ -1,5 +1,4 @@
 import { getConfig } from 'cli';
-import Dispensary from 'dispensary';
 import Linter from 'linter';
 import log from 'logger';
 
@@ -19,7 +18,6 @@ export function createInstance({
 }
 
 export default {
-  Dispensary,
   Linter,
   createInstance,
   isRunFromCLI,
